Move generic indexing logit to the Index ABC, fix the rest
